some groundwork for low-level multi-dim arrays of builtins

+++ src/ProxyWrappers.h
@@ -20,7 +20,7 @@ PyObject* BindCppObjectNoCast(Cppyy::TCppObject_t object,
 PyObject* BindCppObject(Cppyy::TCppObject_t object,
     Cppyy::TCppType_t klass, int flags = 0x0001);
 PyObject* BindCppObjectArray(
-    Cppyy::TCppObject_t address, Cppyy::TCppType_t klass, long* dims);
+    Cppyy::TCppObject_t address, Cppyy::TCppType_t klass, intptr_t* dims);
 
 } // namespace CPyCppyy
